update and add logger's impl

+++ Viso/src/viso/framework/service/protocol/simple/SimpleSgsProtocol.java
@@ -1,7 +1,7 @@
 package viso.framework.service.protocol.simple;
 
 public class SimpleSgsProtocol {
-	public static final int VERSION = 100;
+	public static final byte VERSION = 100;
 	public static final int MAX_MESSAGE_LENGTH = 4096;
 	public static final int MAX_PAYLOAD_LENGTH = 10000;
 	public static final byte SESSION_MESSAGE = 1;
